Add Docker Registry v2 client protocol
Docker Registry V2 Authentication is an OIDC-Like protocol used to authenticate users against a Docker registry.
The Keycloak implementation of this protocol allows for a Keycloak authentication server to be used by a
Docker client to authenticate against a registry. To be able to select this as a protocol when configuring
clients the feature needs to be enabled as it is not available by default. This can be done using the
JAVA_OPTS_APPEND environment and the feature flag for this.